rename `par_iter` module to `iter`
