+++ src/iter/collections.rs
