+++ src/iter/fold.rs
