+++ src/iter/for_each.rs
