+++ src/iter/internal.rs
