+++ src/iter/reduce.rs
