+++ src/option.rs
@@ -2,8 +2,8 @@
 //! (`Option<T>`). You will rarely need to interact with it directly
 //! unless you have need to name one of the iterator types.
 
-use par_iter::*;
-use par_iter::internal::*;
+use iter::*;
+use iter::internal::*;
 use std;
 
 pub struct OptionIter<T: Send> {
