+++ src/prelude.rs
@@ -2,18 +2,16 @@
 //! The intention is that one can include `use rayon::prelude::*` and
 //! have easy access to the various traits and methods you will need.
 
-pub use par_iter::IntoParallelIterator;
-pub use par_iter::IntoParallelRefIterator;
-pub use par_iter::IntoParallelRefMutIterator;
-pub use par_iter::ParallelIterator;
-pub use par_iter::BoundedParallelIterator;
-pub use par_iter::ExactParallelIterator;
-pub use par_iter::IndexedParallelIterator;
-
-pub use par_iter::ToParallelChunks;
-pub use par_iter::ToParallelChunksMut;
-
+pub use iter::BoundedParallelIterator;
+pub use iter::ExactParallelIterator;
+pub use iter::FromParallelIterator;
+pub use iter::IntoParallelIterator;
+pub use iter::IntoParallelRefIterator;
+pub use iter::IntoParallelRefMutIterator;
+pub use iter::IndexedParallelIterator;
+pub use iter::ParallelIterator;
+pub use iter::ToParallelChunks;
+pub use iter::ToParallelChunksMut;
 pub use slice::{SliceIter, ChunksIter, SliceIterMut, ChunksMutIter};
 pub use string::ParallelString;
 
-pub use par_iter::FromParallelIterator;
