+++ src/vec.rs
@@ -2,8 +2,8 @@
 //! (`Vec<T>`). You will rarely need to interact with it directly
 //! unless you have need to name one of those types.
 
-use par_iter::*;
-use par_iter::internal::*;
+use iter::*;
+use iter::internal::*;
 use std;
 
 pub struct VecIter<T: Send> {
